Enables static linking of libcurl
Configures the project to statically link with the libcurl library, resolving potential runtime dependency issues.

This involves adding CURL_STATICLIB to the preprocessor definitions and including the necessary libcurl libraries (libcurl.lib and curlu.lib) as additional dependencies for the AAX, VST2, VST3 and standalone app targets.
Also updates the relative path to libcurl to be project relative instead of user specific.