Fixes macOS build and sandboxing issues.
Addresses issues related to sandboxing on macOS and the inclusion of libcurl.a in the project.

Removes redundant libcurl.a entries in the project file.
Adds multiple libcurl.a references to frameworks build phases to address potential linking issues.

Adds a debug message to the main function to aid in troubleshooting sandboxing related crashes and suggests a terminal command to reset the app's access to the file system.

Removes initialization code related to Superkraft, SK_App_Initializer, SK_Project, and mainWindowHandle, which appears unnecessary at this stage.